Revert "Kaixiang - Create 24/48/72 hours button on team member tasks tab"
